+++ llm/provider/openai/cloudflare_bypass.py
@@ -1,7 +1,10 @@
+import os
 import random
 import time
+from datetime import datetime
 from typing import Optional
 
+import pyautogui
 from DrissionPage import ChromiumOptions, ChromiumPage
 
 from llm import config
@@ -29,7 +32,12 @@ class CloudflareBypass:
         # Some arguments to make the browser better for automation and less detectable.
         if config.NO_GUI:
             logger.info("[CloudflareBypass.__init__] set --no-sandbox")
+            # 最大化窗口保证坐标的准确性
+            arguments.append("--start-maximized")
             arguments.append("--no-sandbox")
+        else:
+            # 图形界面中，如果使用--start-maximized，获取的page_x和page_y有时候并不准确(原因未知)，所以全屏使page_x, page_y为0
+            arguments.append("--start-fullscreen")
 
         for argument in arguments:
             options.set_argument(argument)
@@ -47,6 +55,11 @@ class CloudflareBypass:
                 break
 
             if check_count >= 5:
+                error_time = datetime.now().strftime("%Y-%m-%dT%H:%M:%S-DrissionPage")
+                error_path = os.path.join(config.BROWSER_DATA, "error")
+                self.driver.get_screenshot(
+                    path=error_path, name=f"{error_time}.png", full_page=True
+                )
                 raise Exception("Meet challenge restart")
 
             logger.info(f"Meet challenge and check it. Count: {check_count}")
@@ -59,11 +72,83 @@ class CloudflareBypass:
     def try_to_click_challenge(self):
         try:
             if self.driver.wait.ele_displayed("xpath://div/iframe", timeout=15):
+                if config.env == "dev":
+                    iframe = self.driver("xpath://div/iframe")
+                    iframe = self.driver.get_frame(iframe)
+                    iframe.run_js(
+                        script="""
+                        document.addEventListener('click', function(event) {
+                            console.log("click event trigger")
+                            console.table(event);
+                        });
+                    """
+                    )
                 verify_element = self.driver("xpath://div/iframe").ele(
                     "Verify you are human", timeout=25
                 )
                 time.sleep(random.uniform(2, 5))
-                verify_element.click()
+
+                # 2024-07-05
+                # 直接在element上执行click(通过CDP协议)无法通过cloudflare challenge
+                # 原因:
+                # CDP命令执行的event中client_x, client_y与screen_x, screen_y是一样的，而手动点击触发的事件两者是不一样的,所以无法使用CDP模拟出鼠标点击通过验证
+                # 解决方法:
+                # 先获取点击的坐标，使用pyautogui模拟鼠标点击
+                # CDP参考 https://chromedevtools.github.io/devtools-protocol/tot/Input/
+                # verify_element.click()
+                def generate_biased_random(n):
+                    weights = [min(i, n - i + 1) for i in range(1, n + 1)]
+                    return random.choices(range(1, n + 1), weights=weights)[0]
+
+                if config.env == "dev":
+                    property_list = {
+                        attr: getattr(verify_element.rect, attr)
+                        for attr in dir(verify_element.rect)
+                        if not attr.startswith("__")
+                        and not callable(getattr(verify_element.rect, attr))
+                    }
+
+                    # 手动遍历字典并格式化输出
+                    for key, value in property_list.items():
+                        print(f"{key}: {value}")
+
+                    print("\n")
+                    property_list = {
+                        attr: getattr(self.driver.rect, attr)
+                        for attr in dir(self.driver.rect)
+                        if not attr.startswith("__")
+                        and not callable(getattr(self.driver.rect, attr))
+                    }
+
+                    # 手动遍历字典并格式化输出
+                    for key, value in property_list.items():
+                        print(f"{key}: {value}")
+
+                screen_x, screen_y = verify_element.rect.screen_location
+                page_x, page_y = self.driver.rect.page_location
+                width, height = verify_element.rect.size
+                offset_x, offset_y = generate_biased_random(
+                    int(width - 1)
+                ), generate_biased_random(int(height - 1))
+
+                click_x, click_y = (
+                    screen_x + page_x + offset_x,
+                    screen_y + page_y + offset_y,
+                )
+
+                logger.info(
+                    f"[CloudflareBypass.try_to_click_challenge] Screen point [{screen_x}, {screen_y}]"
+                )
+                logger.info(
+                    f"[CloudflareBypass.try_to_click_challenge] Page point[{page_x}, {page_y}]"
+                )
+                logger.info(
+                    f"[CloudflareBypass.try_to_click_challenge] Click point [{click_x}, {click_y}]"
+                )
+                pyautogui.moveTo(
+                    click_x, click_y, duration=0.5, tween=pyautogui.easeInElastic
+                )
+                pyautogui.click()
                 self.driver.wait.load_start(timeout=20)
         except Exception as e:
             # 2025-05-26